pyproject.toml
[project]
name = "wedding-favor-3d-name"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fonttools>=4.57.0",
    "freetype-py>=2.5.1",
    "lxml>=5.4.0",
    "mapbox-earcut>=1.0.3",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "rtree>=1.4.0",
    "shapely>=2.1.0",
    "svg-path>=6.3",
    "trimesh>=4.6.8",
]
